Add read-from-file and save-JSON functionality

+++ src/main/java/gm/taltech/ee/weatherwise/CurrentWeatherReport.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class CurrentWeatherReport {
-    private Integer temperature;
+    private Double temperature;
     private Integer humidity;
     private Integer pressure;
